refactor(visual-regions): split association helpers
Move visual region geometry, associated text, caption, and label helpers into a focused domain folder so the main builder can stay centered on candidate generation and suppression. This keeps the ~250-line guideline as a cohesion review signal while avoiding behavior changes in the extraction pipeline.
